requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
duckduckgo-search
python-dotenv